requirements.txt
alembic==1.6.5
cachelib==0.1.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-Moment==0.11.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
redis==3.5.3
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3